Re-add most components
